# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
      "AddHookRequest": {
        "additionalProperties": false,
        "properties": {
          "command": {
            "minLength": 1,
            "title": "Command",
            "type": "string"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "event": {
            "minLength": 1,
            "title": "Event",
            "type": "string"
          },
          "id": {
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "match": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match"
          },
          "timeout": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout"
          }
        },
        "required": [
          "id",
          "event",
          "command"
        ],
        "title": "AddHookRequest",
        "type": "object"
      },
...
      "CliMarketplaceDetailResponse": {
        "properties": {
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "githubUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Githuburl"
          },
          "hasMcp": {
            "title": "Hasmcp",
            "type": "boolean"
          },
          "hasSkill": {
            "title": "Hasskill",
            "type": "boolean"
          },
          "iconUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Iconurl"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "installCommand": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Installcommand"
          },
          "isOfficial": {
            "title": "Isofficial",
            "type": "boolean"
          },
          "isTui": {
            "title": "Istui",
            "type": "boolean"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "longDescription": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longdescription"
          },
          "marketplaceUrl": {
            "title": "Marketplaceurl",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "sourceType": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourcetype"
          },
          "stars": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stars"
          },
          "vendorName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vendorname"
          },
          "websiteUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Websiteurl"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "description",
          "marketplaceUrl",
          "hasMcp",
          "hasSkill",
          "isOfficial",
          "isTui"
        ],
        "title": "CliMarketplaceDetailResponse",
        "type": "object"
      },
...
      "DisableHookRequest": {
        "properties": {
          "harness": {
            "description": "Harness identifier",
            "minLength": 1,
            "title": "Harness",
            "type": "string"
          }
        },
        "required": [
          "harness"
        ],
        "title": "DisableHookRequest",
        "type": "object"
      },
...
      "EnableHookRequest": {
        "properties": {
          "harness": {
            "description": "Harness identifier",
            "minLength": 1,
            "title": "Harness",
            "type": "string"
          }
        },
        "required": [
          "harness"
        ],
        "title": "EnableHookRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HookApplyConfigResponse": {
        "properties": {
          "failed": {
            "items": {
              "$ref": "#/components/schemas/HookMutationFailureResponse"
            },
            "title": "Failed",
            "type": "array"
          },
          "hook": {
            "$ref": "#/components/schemas/HookSpecResponse"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "succeeded": {
            "items": {
              "type": "string"
            },
            "title": "Succeeded",
            "type": "array"
          }
        },
        "required": [
          "ok",
          "hook",
          "succeeded",
          "failed"
        ],
        "title": "HookApplyConfigResponse",
        "type": "object"
      },
...
      "HookInventoryEntryResponse": {
        "properties": {
          "canEnable": {
            "title": "Canenable",
            "type": "boolean"
          },
          "displayName": {
            "title": "Displayname",
            "type": "string"
          },
          "enabledStatus": {
            "enum": [
              "enabled",
              "disabled"
            ],
            "title": "Enabledstatus",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "managed",
              "unmanaged"
            ],
            "title": "Kind",
            "type": "string"
          },
          "sightings": {
            "items": {
              "$ref": "#/components/schemas/HookBindingResponse"
            },
            "title": "Sightings",
            "type": "array"
          },
          "spec": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HookSpecResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "displayName",
          "kind",
          "canEnable",
          "enabledStatus",
          "sightings"
        ],
        "title": "HookInventoryEntryResponse",
        "type": "object"
      },
...
      "HookInventoryResponse": {
        "properties": {
          "columns": {
            "items": {
              "$ref": "#/components/schemas/HookInventoryColumnResponse"
            },
            "title": "Columns",
            "type": "array"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/HookInventoryEntryResponse"
            },
            "title": "Entries",
            "type": "array"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/HookInventoryIssueResponse"
            },
            "title": "Issues",
            "type": "array"
          }
        },
        "required": [
          "columns",
          "entries"
        ],
        "title": "HookInventoryResponse",
        "type": "object"
      },
...
      "HookMutationResponse": {
        "properties": {
          "hook": {
            "$ref": "#/components/schemas/HookSpecResponse"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "hook"
        ],
        "title": "HookMutationResponse",
        "type": "object"
      },
      "HookSetHarnessesResultResponse": {
        "properties": {
          "failed": {
            "items": {
              "$ref": "#/components/schemas/HookMutationFailureResponse"
            },
            "title": "Failed",
            "type": "array"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "succeeded": {
            "items": {
              "type": "string"
            },
            "title": "Succeeded",
            "type": "array"
          }
        },
        "required": [
          "ok",
          "succeeded",
          "failed"
        ],
        "title": "HookSetHarnessesResultResponse",
        "type": "object"
      },
...
      "OkResponse": {
        "properties": {
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "OkResponse",
        "type": "object"
      },
...
      "PromoteHookRequest": {
        "additionalProperties": false,
        "properties": {
          "observedHarness": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed harness"
          }
        },
        "title": "PromoteHookRequest",
        "type": "object"
      },
...
      "ReconcileHookRequest": {
        "additionalProperties": false,
        "properties": {
          "harnesses": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Harnesses"
          },
          "observedHarness": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed harness"
          },
          "sourceKind": {
            "enum": [
              "managed",
              "harness"
            ],
            "title": "Sourcekind",
            "type": "string"
          }
        },
        "required": [
          "sourceKind"
        ],
        "title": "ReconcileHookRequest",
        "type": "object"
      },
...
      "SetHookHarnessesRequest": {
        "properties": {
          "target": {
            "enum": [
              "enabled",
              "disabled"
            ],
            "title": "Target",
            "type": "string"
          }
        },
        "required": [
          "target"
        ],
        "title": "SetHookHarnessesRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "harness-asset-manager",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/hooks": {
      "get": {
        "operationId": "list_hooks_api_hooks_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HookInventoryResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Hooks"
      },
      "post": {
        "operationId": "create_hook_api_hooks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddHookRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HookMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Hook"
      }
    },
    "/api/hooks/{id}": {
      "delete": {
        "operationId": "delete_hook_api_hooks__id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HookSetHarnessesResultResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Hook"
      },
      "get": {
        "operationId": "get_hook_api_hooks__id__get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HookInventoryEntryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Hook"
      }
    },
    "/api/hooks/{id}/disable": {
      "post": {
        "operationId": "disable_hook_api_hooks__id__disable_post",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisableHookRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Disable Hook"
      }
    },
    "/api/hooks/{id}/enable": {
      "post": {
        "operationId": "enable_hook_api_hooks__id__enable_post",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnableHookRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enable Hook"
      }
    },
    "/api/hooks/{id}/promote": {
      "post": {
        "operationId": "promote_hook_api_hooks__id__promote_post",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromoteHookRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HookMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Promote Hook"
      }
    },
    "/api/hooks/{id}/reconcile": {
      "post": {
        "operationId": "reconcile_hook_api_hooks__id__reconcile_post",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReconcileHookRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HookApplyConfigResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reconcile Hook"
      }
    },
    "/api/hooks/{id}/set-harnesses": {
      "post": {
        "operationId": "set_hook_harnesses_api_hooks__id__set_harnesses_post",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetHookHarnessesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HookSetHarnessesResultResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Hook Harnesses"
      }
    },
    "/api/marketplace/clis/items/{slug}": {
      "get": {
        "operationId": "get_cli_marketplace_detail_api_marketplace_clis_items__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CliMarketplaceDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Cli Marketplace Detail"
      }
    },
    "/api/marketplace/clis/popular": {
      "get": {
        "operationId": "popular_cli_marketplace_api_marketplace_clis_popular_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CliMarketplacePageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Popular Cli Marketplace"
      }
    },
    "/api/marketplace/clis/search": {
      "get": {
        "operationId": "search_cli_marketplace_api_marketplace_clis_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CliMarketplacePageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Cli Marketplace"
      }
    },
...
    "/api/marketplace/items/{item_id}": {
      "get": {
        "operationId": "get_marketplace_detail_api_marketplace_items__item_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Marketplace Detail Api Marketplace Items  Item Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Marketplace Detail"
      }
    },
...
    "/api/marketplace/popular": {
      "get": {
        "operationId": "popular_marketplace_api_marketplace_popular_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Popular Marketplace Api Marketplace Popular Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Popular Marketplace"
      }
    },
    "/api/marketplace/search": {
      "get": {
        "operationId": "search_marketplace_api_marketplace_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Search Marketplace Api Marketplace Search Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Marketplace"
      }
    },
...
  }
}